refactor test mocks
